server/requirements.txt
asgiref==3.5.0
boto3==1.21.37
botocore==1.24.37
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.2
decorator==5.1.1
defusedxml==0.7.1
Django==3.2.5
django-cors-headers==3.7.0
django-filter==21.1
django-jazzmin==2.4.8
django-storages==1.12.3
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-yasg==1.20.0
future==0.18.2
geocoder==1.38.1
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.1
jmespath==1.0.0
MarkupSafe==2.1.1
oauthlib==3.2.0
packaging==21.3
psycopg2-binary==2.9.1
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.7
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.1
ratelim==0.1.6
requests==2.27.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
s3transfer==0.5.2
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.9
whitenoise==5.3.0
